factor out cline rules functionality
